Revert "Move all SelectField option lists into a shared SelectorOptions const"
This reverts commit 3bcf948ac9fe6b221b138759f159b7154f0ac2fc.
