+++ src/components/items/types/devices/forms/programFormFields.tsx
@@ -2,7 +2,15 @@ import Stack from "@mui/material/Stack"
 
 import { withFieldGroup } from "#/integrations/tanstackForm/useAppForm.ts"
 import { programFormOpts } from "#/lib/hooks/items/types/devices/forms/useProgramForm.tsx"
-import { SelectorOptions } from "#/system/selectorOptions.tsx"
+import { ProgramType } from "#/system/gear/programData.ts"
+
+const splitCamelCase = (s: string) => s.replace(/([a-z0-9])([A-Z])/g, "$1 $2")
+const titleCase = (s: string) => s.split(" ").map((w) => w.charAt(0).toUpperCase() + w.slice(1)).join(" ")
+
+const programTypeOptions = Object.values(ProgramType).map((programType) => ({
+  label: titleCase(splitCamelCase(programType)),
+  value: programType,
+}))
 
 export const ProgramFormFields = withFieldGroup({
   ...programFormOpts,
@@ -15,7 +23,7 @@ export const ProgramFormFields = withFieldGroup({
               label="Program Type"
               fullWidth
               size="small"
-              options={SelectorOptions.programType}
+              options={programTypeOptions}
             />
           )}
         </group.AppField>
