+++ src/components/items/types/weapons/forms/weaponSkillFormFields.tsx
@@ -2,7 +2,23 @@ import Stack from "@mui/material/Stack"
 
 import { withFieldGroup } from "#/integrations/tanstackForm/useAppForm.ts"
 import { weaponFormOpts } from "#/lib/hooks/items/types/weapons/forms/useWeaponForm.tsx"
-import { SelectorOptions } from "#/system/selectorOptions.tsx"
+import { AttributeKey } from "#/system/attributeKey.ts"
+import { skillList } from "#/system/skills/skillList.ts"
+
+const skillOptions = Object.entries(skillList)
+  .filter(([_, skill]) => skill.isWeaponSkill)
+  .map(([key, _]) => ({
+    label: key,
+    value: key,
+  }))
+
+const attributeOptions = [
+  { label: "None", value: "" },
+  ...Object.values(AttributeKey).map((key) => ({
+    label: key.charAt(0).toUpperCase() + key.slice(1),
+    value: key,
+  })),
+]
 
 export const WeaponSkillFormFields = withFieldGroup({
   ...weaponFormOpts,
@@ -15,7 +31,7 @@ export const WeaponSkillFormFields = withFieldGroup({
               label="Skill"
               size="small"
               sx={{ flex: 1 }}
-              options={SelectorOptions.weaponSkill}
+              options={skillOptions}
             />
           )}
         </group.AppField>
@@ -26,7 +42,7 @@ export const WeaponSkillFormFields = withFieldGroup({
               label="Attribute"
               size="small"
               sx={{ flex: 1 }}
-              options={SelectorOptions.weaponSkillAttribute}
+              options={attributeOptions}
             />
           )}
         </group.AppField>
