+++ src/components/runner/magician/spells/form/spellFormFields.tsx
@@ -8,8 +8,14 @@ import { GameEffectsFieldGroup } from "#/components/system/gameEffects/gameEffec
 import { SourceFieldGroup } from "#/components/system/sources/sourceFieldGroup.tsx"
 import { Label } from "#/components/ui/text/label.tsx"
 import type { SpellForm } from "#/lib/hooks/runner/magician/spells/form/useSpellForm.ts"
-import { SpellDrainType } from "#/system/magic/spellData.ts"
-import { SelectorOptions } from "#/system/selectorOptions.tsx"
+import {
+  SpellCategory,
+  SpellDamage,
+  SpellDrainType,
+  SpellDuration,
+  SpellRange,
+  SpellType,
+} from "#/system/magic/spellData.ts"
 
 interface SpellFormFieldsProps {
   form: SpellForm
@@ -27,7 +33,7 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
           {(field) => (
             <field.SelectField
               label="Category"
-              options={SelectorOptions.spellCategory}
+              options={Object.values(SpellCategory).map((value) => ({ label: value, value }))}
             />
           )}
         </form.AppField>
@@ -37,7 +43,7 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
             <field.SelectField
               label="Type"
               sx={{ flexGrow: 1 }}
-              options={SelectorOptions.spellType}
+              options={Object.values(SpellType).map((value) => ({ label: value, value }))}
             />
           )}
         </form.AppField>
@@ -46,7 +52,11 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
           {(field) => (
             <field.SelectField
               label="Range"
-              options={SelectorOptions.spellRange}
+              options={[
+                { label: "Touch", value: SpellRange.Touch },
+                { label: "Line of Sight", value: SpellRange.LoS },
+                { label: "Line of Sight (Area)", value: SpellRange.LoSArea },
+              ]}
             />
           )}
         </form.AppField>
@@ -60,7 +70,7 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
             <field.SelectField
               label="Duration"
               sx={{ flexGrow: 1 }}
-              options={SelectorOptions.spellDuration}
+              options={Object.values(SpellDuration).map((value) => ({ label: value, value }))}
             />
           )}
         </form.AppField>
@@ -71,7 +81,10 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
               <field.SelectField
                 label="Drain"
                 sx={{ flexGrow: 1 }}
-                options={SelectorOptions.spellDrainType}
+                options={[
+                  { label: "Force ÷ 2", value: SpellDrainType.Force },
+                  { label: "Fixed", value: SpellDrainType.Fixed },
+                ]}
               />
             )}
           </form.AppField>
@@ -125,7 +138,7 @@ export const SpellFormFields: FC<SpellFormFieldsProps> = ({ form }) => {
                       disabled={!dealsDmgfield.state.value}
                       label="Damage Type"
                       sx={{ flexGrow: 1 }}
-                      options={SelectorOptions.spellDamage}
+                      options={Object.values(SpellDamage).map((value) => ({ label: value, value }))}
                     />
                   )}
                 </form.AppField>
