+++ src/system/sourceData.ts
@@ -49,6 +49,12 @@ export const books = {
 
 export type BookKey = keyof typeof books
 
+/**
+ * List of available books and their display labels.
+ */
+export const bookOptions = Object.entries(books)
+  .map(([key, value]) => ({ value: key as BookKey, ...value }))
+
 export const formatBookRef = (source: SourceData) => {
   return `${source.book} p.${source.page}`
 }
